Updated workshop with the generic extension method.

+++ public/step_05/solution.dart
@@ -1,10 +1,8 @@
 main() {
   int counter = 1.incrementCounter();
   print('counter: $counter');
-  // Compile time error
-  // A value of type 'double' can't be assigned to a variable of type 'int'.
-  // TODO(1) Try solving this compile time error
-  int shouldBeDouble = 2.0.incrementCounter();
+
+  double shouldBeDouble = 2.0.incrementCounter();
   print('counter: $shouldBeDouble');
 }
